pyproject.toml
[tool.poetry]
name = "libraryapi"
version = "0.1.0"
description = ""
authors = ["Hordunlarmy <Hordunlarmy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.114.2"
uvicorn = "^0.30.6"
flask = "^3.0.3"
python-decouple = "^3.8"
gunicorn = "^23.0.0"
pymongo = "^4.8.0"
mysql-connector-python = "^9.0.0"
pika = "^1.3.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
isort = "^5.13.2"
autopep8 = "^2.3.1"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
